shouldWork
