+++ superlists/carpool/views.py
@@ -98,9 +98,5 @@ def find_riders_for_a_driver(user):
     if (debugging):
         return Rider.objects.all()
     else:
-        filtered_riders = Rider.objects.filter(date = user.date
-                                    ).filter(start__iexact = user.start
-                                    ).filter(end__iexact = user.end)[:5]                           
-        for item in filtered_riders:
-            print (item.nameFirst + "Hello\n")
+        filtered_riders = Rider.objects.filter(date = user.date) #.filter(end__iexact = user.end)[:5]                           
         return filtered_riders
